feat: Enhance tagging process in versioning tool
The changes implemented enhance the tagging system in the versioning tool, adding notification message output for tagging and pushing. It maintains the basic functionalities but adds steps to not only create a tag for the current commit but also pushes the tag to the repository. In addition, an updated version of the tool has been used.